requirements.txt
streamlit
langchain-community
faiss-cpu
sentence-transformers
cerebras-cloud-sdk
